pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page loader"
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
]
authors = ["Nella <nella611@yandex.ru>"]
packages = [
  { include = "page_loader" },
]


[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"
python-slugify = "^5.0.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
requests-mock = "^1.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
page-loader = "page_loader.scripts.page_loader:main"
